use the correct type
git-svn-id: https://svn.participatoryculture.org/svn/channelguide2/trunk@1124 be7adf91-e322-0410-8f47-e6edb61c52aa

+++ channelguide/cache/middleware.py
@@ -47,7 +47,7 @@ class CacheMiddlewareBase(object):
             namespace_names = self.namespace
         namespace_values = client.get_multi(namespace_names)
         for name in namespace_names:
-            if type(namespace_values.get(name, None)) is not int:
+            if type(namespace_values.get(name, None)) is not float:
                 namespace_values[name] = None
             if namespace_values.get(name, None) is None:
                 namespace_values[name] = time.time()
